Cleanup old schema code and refactoring
