+++ src/main/java/uk/gov/indexer/Indexer.java
@@ -4,9 +4,8 @@ import com.google.common.base.Throwables;
 import org.skife.jdbi.v2.DBI;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import uk.gov.indexer.dao.DestinationDBUpdateDAO;
 import uk.gov.indexer.dao.DestinationDBUpdateDAO_NewSchema;
-import uk.gov.indexer.dao.IndexedEntriesUpdateDAO;
+import uk.gov.indexer.dao.EntryUpdateDAO;
 import uk.gov.indexer.dao.SourceDBQueryDAO;
 import uk.gov.indexer.monitoring.CloudwatchRecordsProcessedUpdater;
 
@@ -18,7 +17,6 @@ public class Indexer {
 
     private final Configuration configuration;
     private final String register;
-    private DestinationDBUpdateDAO destinationDBUpdateDAO;
     private DestinationDBUpdateDAO_NewSchema destinationDBUpdateDAO_NewSchema;
     private SourceDBQueryDAO sourceDBQueryDAO;
 
@@ -32,27 +30,14 @@ public class Indexer {
             LOGGER.info("setting up register " + register);
 
             DBI destDbi = new DBI(configuration.getProperty(register + ".destination.postgres.db.connectionString"));
-            destinationDBUpdateDAO = destDbi.open().attach(DestinationDBUpdateDAO.class);
             destinationDBUpdateDAO_NewSchema = destDbi.open().attach(DestinationDBUpdateDAO_NewSchema.class);
 
             DBI sourceDbi = new DBI(configuration.getProperty(register + ".source.postgres.db.connectionString"));
             sourceDBQueryDAO = sourceDbi.onDemand(SourceDBQueryDAO.class);
 
-
-            executorService.scheduleAtFixedRate(
-                    new IndexerTask(
-                            configuration.cloudwatchEnvironmentName().map(e -> new CloudwatchRecordsProcessedUpdater(e, register)),
-                            register,
-                            sourceDBQueryDAO,
-                            destinationDBUpdateDAO
-                    ),
-                    0,
-                    10,
-                    TimeUnit.SECONDS
-            );
-
             executorService.scheduleAtFixedRate(
                     new IndexEntryItemTask(
+                            configuration.cloudwatchEnvironmentName().map(e -> new CloudwatchRecordsProcessedUpdater(e, register)),
                             register,
                             sourceDBQueryDAO,
                             destinationDBUpdateDAO_NewSchema
@@ -64,22 +49,12 @@ public class Indexer {
 
             configuration.cloudSearchEndPoint(register).ifPresent(
                     endPoint -> executorService.scheduleAtFixedRate(
-                            new CloudSearchDataUploadTask(register, endPoint, configuration.cloudSearchWaterMarkEndPoint(register).get(), destDbi.onDemand(IndexedEntriesUpdateDAO.class)),
+                            new CloudSearchDataUploadTask(register, endPoint, configuration.cloudSearchWaterMarkEndPoint(register).get(), destDbi.onDemand(EntryUpdateDAO.class)),
                             0,
                             5,
                             TimeUnit.MINUTES
                     )
             );
-
-            configuration.elasticSearchEndPoint(register).ifPresent(
-                    endPoint -> executorService.scheduleAtFixedRate(
-                            new ElasticSearchDataUploadTask(register, endPoint, destDbi.onDemand(IndexedEntriesUpdateDAO.class)),
-                            0,
-                            5,
-                            TimeUnit.MINUTES
-                    )
-            );
-
         } catch (Throwable e) {
             LOGGER.info("Error occurred while setting indexer for register: " + register + ". Error is -> " + e.getMessage());
             LOGGER.error(Throwables.getStackTraceAsString(e));
@@ -88,9 +63,6 @@ public class Indexer {
     }
 
     public synchronized void stop() {
-        if (destinationDBUpdateDAO != null) {
-            destinationDBUpdateDAO.close();
-        }
         if (destinationDBUpdateDAO_NewSchema != null) {
             destinationDBUpdateDAO_NewSchema.close();
         }
