+++ src/main/java/uk/gov/indexer/dao/EntryUpdateDAO.java
@@ -1,20 +1,27 @@
 package uk.gov.indexer.dao;
 
-import org.skife.jdbi.v2.sqlobject.BindBean;
-import org.skife.jdbi.v2.sqlobject.SqlBatch;
-import org.skife.jdbi.v2.sqlobject.SqlQuery;
-import org.skife.jdbi.v2.sqlobject.SqlUpdate;
+import org.skife.jdbi.v2.sqlobject.*;
+import org.skife.jdbi.v2.sqlobject.customizers.RegisterMapper;
+import org.skife.jdbi.v2.sqlobject.stringtemplate.UseStringTemplate3StatementLocator;
 
+import java.util.List;
+
+@UseStringTemplate3StatementLocator("/sql/init_entry.sql")
 public interface EntryUpdateDAO extends DBConnectionDAO {
 
-    String ENTRY_TABLE = "entry";
+    @SqlUpdate
+    void ensureEntrySchemaInPlace();
 
-    @SqlUpdate("create table if not exists entry (entry_number SERIAL PRIMARY KEY, sha256hex VARCHAR, timestamp timestamp)")
-    void ensureEntryTableInPlace();
-
-    @SqlQuery("SELECT MAX(entry_number) FROM " + ENTRY_TABLE)
+    @SqlQuery("SELECT MAX(entry_number) FROM entry")
     int lastReadEntryNumber();
 
-    @SqlBatch("INSERT INTO " + ENTRY_TABLE + "(entry_number, sha256hex, timestamp) VALUES(:entryNumber, :itemHash, :timestamp)")
+    @SqlBatch("INSERT INTO entry(entry_number, sha256hex, timestamp) VALUES(:entryNumber, :itemHash, :timestamp)")
     void writeBatch(@BindBean Iterable<Entry> entries);
+
+    @SqlUpdate("update total_entries set count=count+:entriesWritten")
+    void updateTotalEntries(@Bind("entriesWritten") long entriesWritten);
+
+    @RegisterMapper(RecordMapper.class)
+    @SqlQuery("SELECT entry_number, entry.sha256hex as sha256hex, timestamp, content FROM item, entry WHERE item.sha256hex=entry.sha256hex and entry_number > :lastReadEntryNumber ORDER BY entry_number LIMIT 5000")
+    List<Record> fetchRecordsAfter(@Bind("lastReadEntryNumber") int watermark);
 }
