+++ src/main/resources/application.properties
@@ -1,10 +1,8 @@
 address.source.postgres.db.connectionString=jdbc:postgresql://localhost:5432/address?user=postgres
-address.destination.postgres.db.connectionString=jdbc:postgresql://localhost:5432/address?user=postgres
+address.destination.postgres.db.connectionString=jdbc:postgresql://localhost:5432/address_p?user=postgres
 
-address.cloudsearch.search.endpoint=ConfigureSearchDomainEndPointForTheRegister
-address.cloudsearch.highwatermark.endpoint=ConfigureHighwatermarkEndPointForRegister
-
-address.elasticsearch.endpoint=ConfigureElasticSearchEndpointUrl_must_start_with_http_or_https
+#address.cloudsearch.search.endpoint=ConfigureSearchDomainEndPointForTheRegister
+#address.cloudsearch.highwatermark.endpoint=ConfigureHighwatermarkEndPointForRegister
 
 # If set, CloudWatch monitoring is enabled and this specifies the CloudWatch namespace
 #cloudwatch.environment.name=local
